refactor: modal에 dynamic import 적용
