+++ src/view/onboarding/Section/Section1.tsx
@@ -24,18 +24,16 @@ const intro = { section: "intro" } as const;
 const Section1 = () => {
   const router = useRouter();
   const handleClick = () => {
-    if (typeof document !== "undefined") {
-      const featureSection = document.querySelector("#feature-1");
-      featureSection?.scrollIntoView({
-        behavior: "smooth",
-      });
-    }
+    const featureSection = document.querySelector("#feature-1");
+    featureSection?.scrollIntoView({
+      behavior: "smooth",
+    });
   };
   const handleLoginClick = () => {
-    router.push("/login");
+    router.push("/login", { scroll: false });
   };
   const handleSignupClick = () => {
-    router.push("/signup");
+    router.push("/signup", { scroll: false });
   };
   return (
     <section
